chore: improve build for both envs
