+++ src/production/engine.ts
@@ -12,7 +12,7 @@ export function Engine(productionContext: IProductionContext) {
 
       for (const phase of phases) {
         const name = phase.toString().match(/function (\w+)/i)[1];
-        productionContext.log.info('phase', 'Running ' + name);
+        productionContext.log.info('phase', `Running ${name}`);
         await phase(productionContext);
       }
     },
