+++ src/production/phases/BundlePhase.ts
@@ -1,5 +1,4 @@
-import { createBundleRouter } from '../../bundle/bundleRouter';
-import { createRunResponse } from '../../core/runResponse';
+import { createBuild } from '../../core/build';
 import { createRuntimeRequireStatement } from '../../moduleResolver/moduleResolver';
 import { IProductionContext } from '../ProductionContext';
 
@@ -25,26 +24,12 @@ export async function BundlePhase(productionContext: IProductionContext) {
   }
 
   if (modules) {
-    const router = createBundleRouter({ ctx, entries });
-    router.generateBundles(modules);
-    if (productionContext.splitEntries.entries.length > 0) {
-      router.generateSplitBundles(productionContext.splitEntries.entries);
-    }
-    await ctx.ict.resolve();
-
-    const { bundles, manifest, onComplete } = await createRunResponse(ctx, router);
-
-    productionContext.runResponse = {
+    productionContext.runResponse = await createBuild({
       bundleContext: productionContext.bundleContext,
-      bundles,
       entries,
-      manifest,
       modules: productionContext.modules,
-      onComplete,
       splitEntries: productionContext.splitEntries,
-    };
-
-    ctx.log.stopStreaming();
-    ctx.ict.sync('complete', productionContext.runResponse);
+      ctx,
+    });
   }
 }
